+++ insta-clone/src/Components/Login/Login.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Login = () => {
+    return (
+        <div>here's the login</div>
+    )
+}
+
+export default Login;
